By default include all package data.
This more closely matches the flit/poetry behavior, but does not yet
handle any excludes.  With this changes, on the top 5000 packages:

```
2100 dowsing computes the same* source mapping as the wheel
579  dowsing computes a different* source mapping than the wheel
327  dowsing thinks there are no sources
145  dowsing raises an exception (some of these are 404)

1849 do not have wheel+sdist to easily check
```

`*` -- this is after some simple normalization, but many of these are
either harmless differences, or exposing other problems.  For example:

`pytz` includes the in-package tests w/ dowsing

`boto3` wheel cannot be recreated from the published sdist

`inflection` wheel includes both a module and a package, and cannot be
recreated from the published sdist

`bleach` includes the dist-info for a vendored dep

`regex` produces filenames with `.` instead of `/` because of an unusual
use of `py_modules` containing dots.