More robust evaluation for self-referential names
Tracks the `target_name` that we are recursively evaluating, and
short-circuits evaluation if we attempt to evaluate that name again.
Also allows the assignment evaluation to try multiple assignments until
either a real value is found, or all assignments are exhausted. Also
prevents combining lhs and rhs of binary addition if either side is the
`"??"` sentinel value.

Fixes #56